# OpenAPI/customer-facade API.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "Customer" : {
        "type" : "object",
        "properties" : {
          "cpr" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "customerBankId" : {
            "type" : "string"
          }
        }
      }
    }
  },
  "paths" : {
    "/customer" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Customer"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Register Customer",
        "tags" : [ "Customer Facade Resource" ]
      }
    },
    "/customer/{customerId}" : {
      "delete" : {
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Deregister Customer",
        "tags" : [ "Customer Facade Resource" ]
      }
    },
    "/customer/{customerId}/report" : {
      "get" : {
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get Customer Report",
        "tags" : [ "Customer Facade Resource" ]
      }
    },
    "/customer/{customerId}/report/history" : {
      "get" : {
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get Customer Report History",
        "tags" : [ "Customer Facade Resource" ]
      }
    },
    "/customer/{customerId}/tokens" : {
      "put" : {
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "count",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get Tokens",
        "tags" : [ "Customer Facade Resource" ]
      }
    }
  },
  "info" : {
    "title" : "customer-facade API",
    "version" : "1.0.0"
  }
}